+++ horde/apis/v2/stable.py
@@ -259,16 +259,19 @@ class ImageAsyncGenerate(GenerateTemplate):
             )
         if needs_kudos is True:
             if required_kudos > self.user.kudos:
-                self.wp.delete()
-                raise e.KudosUpfront(
-                    required_kudos,
-                    self.username,
-                    message=f"Due to heavy demand, for requests over {resolution}x{resolution} "
-                    "or over 50 steps (10 steps for LCM work, 30 steps for Stable Cascade, "
-                    "and half those steps for k_heun, dpmpp_sde, and dpm_2*) "
-                    "the client needs to already have the required kudos. "
-                    f"This request requires {required_kudos} kudos to fulfil.",
-                )
+                if self.args.allow_downgrade:
+                    self.wp.downgrade(resolution)
+                else:
+                    self.wp.delete()
+                    raise e.KudosUpfront(
+                        required_kudos,
+                        self.username,
+                        message=f"Due to heavy demand, for requests over {resolution}x{resolution} "
+                        "or over 50 steps (10 steps for LCM work, 30 steps for Stable Cascade, "
+                        "and half those steps for k_heun, dpmpp_sde, and dpm_2*) "
+                        "the client needs to already have the required kudos. "
+                        f"This request requires {required_kudos} kudos to fulfil.",
+                    )
         if self.wp.params["steps"] >= 300:
             print_args = self.args.copy()
             print_args["apikey"] = "REDACTED"
